FIX: filepath for macOS

+++ git/setup_git.sh
@@ -25,4 +25,16 @@ git config --global pull.rebase true
 git config --global fetch.prune true
 git config --list
 
-mkdir -p ~/.config/git && cp /home/"$TARGET_USER"/dotfiles/git/.gitignore_global /home/"$TARGET_USER"/.config/git/ignore
+# set $OS
+if [ "$(uname)" = 'Darwin' ]; then
+	OS='Mac'
+	mkdir -p ~/.config/git && cp /Users/"$TARGET_USER"/dotfiles/git/.gitignore_global /Users/"$TARGET_USER"/.config/git/ignore
+elif [ "$(uname -s | cut -c1-5)" = 'Linux' ]; then
+	OS='Linux'
+	mkdir -p ~/.config/git && cp /home/"$TARGET_USER"/dotfiles/git/.gitignore_global /home/"$TARGET_USER"/.config/git/ignore
+elif [ "$(uname -s | cut -c1-10)" = 'MINGW32_NT' ]; then
+	OS='Cygwin'
+else
+	echo -e "${FAILED}Your platform ($(uname -a)) is not supported."
+	exit 1
+fi
